// src/app/features/expenses/expense-type/models/expense-type-dto.model.ts
export interface ExpenseTypeDto {
    id: number;
    name: string;
    description: string;
    isActive: boolean;
    isDeleted: boolean;
    createdBy: number | null;
    updatedBy: number | null;
    deletedBy: number | null;
    createdAt: Date | null;
    updatedAt: Date | null;
    deletedAt: Date | null;
  }
  